requirements.txt
loguru
firebase-admin
python-telegram-bot
moviepy